move to evm.rs and new helper
